Install all security updates
The playbook updates a few specific packages to their latest version.
At least in the case of bash, this was done in response to a specific
vulnerability (CVE-2014-6271); but the Ubuntu cloud image is re-built
frequently and it no longer contains an affected version. This does
not handle other vulnerabilities as they arise: at the time of this
writing, there are security updates for the login and passwd packages
contained in the latest Ubuntu cloud image, but they do not get fixed
this way.

A better approach is to apply any updates to installed packages which
are found in the "security" repository.