Make usage of unofficial PPA for newer Python 3 versions configurable
Commit 259bb94 added the "New Python Versions" PPA from Felix Krull
when using this playbook. The PPA description says:

    Disclaimer: there's no guarantee of timely updates in case of
    security problems or other issues. If you want to use them in a
    security-or-otherwise-critical environment (say, on a production
    server), you do so at your own risk.

Because of this, it would be better for those who wish to use this
PPA to explicitly enable it; add a variable for this. Ubuntu 14.04
already includes Python 3.4 in the main repository, which receives
updates from the Ubuntu Security Team.

When using Python from the main repository, suggest specifying the
version using the alias "python" or "python3" (although the explicit
version "python2.7" or "python3.4" can still be given). This makes
the setting forward-compatible with Ubuntu 16.04, which includes a
different version of Python 3.