Ensure libpq-dev is installed from the PostgreSQL repository
The PostgreSQL apt repository was configured in the db role. This
role is applied after the base role, where libpq-dev was installed.
This package has different versions in the main apt repository and
the PostgreSQL apt repository.

Because libpq-dev is only used in the web role, install it there.
Move the PostgreSQL apt repository configuration to the base role
to ensure that libpq-dev is installed from it.

Consider that the dbservers.yml and webservers.yml playbooks apply
the db role and the web role on separate hosts, but they both need
to install packages from the PostgreSQL apt repository.