Support distribution versions with Python 3 as default interpreter
Ubuntu 16.04 uses Python 3 by default. Ansible 2.2 and later can be
configured to use Python 3 as the interpreter, which eliminates the
need to install Python 2.

Install the appropriate packages for pip, ipython, and psycopg2 in
either case. Ensure that the default Python version used in the
virtualenv is the default Python version used in the distribution,
even if that is Python 3.