Improve information about Python 3 in README.md
Users should not normally need to change the Python interpreter used
by Ansible, even when using Ubuntu 14.04 LTS: the "minimal core" for
Ubuntu 14.04 LTS installations includes Python 3. (This can be seen
by examining in the dependencies for the "ubuntu-minimal" package.)
As a result Python 3 is installed in the cloud images and the Docker
container. (Note that Ubuntu 12.04 LTS has now reached End of Life.)

The main package repositories for Ubuntu 14.04/16.04 LTS contain an
older version of Ansible that does not support using Python 3 as the
interpreter on remote machines. Provide instructions to enable the
official Ansible PPA on host systems using one of these releases, in
order to install a more recent version of Ansible.

Rearrange other text in README.md for clarity.