Install Avahi in VM to enable mDNS
If mDNS is enabled in the VM, it will advertise a local domain name on
the private network that can be used to connect to it. For example, if
the hostname of the VM is "my-cool-app", then it can be accessed using
the domain name "my-cool-app.local" from a web browser or ssh client on
the host system. (This requires mDNS to be enabled on the host system
as well; this is usually the case out-of-the-box on OS X and Linux).

Enable mDNS in the VM, by setting its hostname and installing the Avahi
daemon.

In the Vagrantfile, define the VM using the mDNS domain name. This name
will then be used for the hostname in the generated Ansible inventory
file (and as a result, for the value of the inventory_hostname variable
when processing the Ansible playbooks). Otherwise, Vagrant will use the
hostname "default" in the Ansible inventory.

Replace the demo self-signed certificate, using the mDNS domain name as
the subject, rather than *.example.com. (The private key is unchanged.)