Add Vagrant configuration for the Docker provider
Docker can be used on a Linux system to run this application inside
a Linux container (a lightweight alternative to a virtual machine).

This adds the configuration for Vagrant to use Docker as a provider.
It includes a Dockerfile to build a container image that works as a
Vagrant base box (https://www.vagrantup.com/docs/boxes/base.html).
To allow the container to be provisioned by Ansible in the existing
way used by other providers, it runs systemd as its entrypoint; see
https://developers.redhat.com/blog/2016/09/13/running-systemd-in-a-non-privileged-container/
for more background. Certain hosts may require extra configuration
before systemd can be run by Docker; in Fedora, the SELinux boolean
container_manage_cgroup must be enabled, for example.

This change does not make Docker the default provider for Vagrant.
To use Docker, either call "vagrant up --provider=docker", or set
the environment variable "VAGRANT_DEFAULT_PROVIDER=docker".