Improved support for different project layouts.
The current project layout assumes manage.py wsgi.py and the requirements
file can be found in the root directory of the project. You can override
this by editing paths in roles/web/defaults/main.yml but this means digging
around in the internals of the project. It would be much nice if all the
user editable parts were restricted to the env_var directory.

The solution was to a "Project layout" section in env_vars/base.yml where
the paths, relative to the root of the project can be defined. Since the
requirements are often environment specific the relative path was also
defined in env_vars/vagrant.yml and env_vars/development.yml so it can
be overridden as needed.

In roles/web/defaults/main.yml the definition for reuirements_file was
changed to use the relative path and a new variable, django_path, was
added, allowing the location of manage.py and wsgi.py to be decoupled
from project_path. This gives more flexibility if the user still needs
to go in and edit this file to match their project layout.

NOTE: the changes to roles/web/defaults/main.yml are sufficient to support
different layouts but creating the relative paths is useful so only the
env_vars files need to be edited.