Upgrade base image to Bionic Beaver
- Ubuntu 14.04 is rapidly approaching EOL (April, 2019).
- Python 2 will no longer be supported in one year's time (Jan, 2020)

In this day and age, nobody should really be running infrastructure on
32-bit machines.

Update the README to refer to 18.04 as the default image, with 16.04
supported as well. Python 3 is the default version as well - Python 2
can still be supported, but anybody starting out with this repo likely
will not be deploying Python 2: tune instructions to assume that 18.04
with Python 3 is the default starting point for everyone.